backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
google-generativeai
python-dotenv
requests
aiofiles
pillow
